refactor: [#7] extract module
